GET Foods moved to Service
