+++ lib/index.js
@@ -39,22 +39,9 @@ function makeMealsPage() {
 }
 
 
+import FoodService from './services/food_service.js'
+
 function makeFoodsPage() {
-  clearContainer()
-  const foodTable = new FoodTable()
-  let element     = foodTable.new_section()
-  let container   = foodTable.make_specific_table(element)
-  var target = 'https://protected-retreat-87261.herokuapp.com/api/v1/foods'
-  get_service(foodTable, target)
-  function get_service(model, target) {
-    fetch(target)
-      .then(blob => blob.json() )
-      .then(data => {
-        model.make_table_rows(data)
-      })
-      .catch(e => {
-        console.log(e);
-        return e;
-      });
-  }
+  let service = new FoodService
+  service.makeFoodsPage()
 }
